Handle zero-length dictionary keys.
Rather than using an empty string ('') as the key argument to add_item
to indicate there is no key, use the value None instead.  This allows
an actual key of "" to be displayed properly.

Also show the keys in quotation marks even when the value is a dict or array.